add (_count_ 0) optional variable.
This right now can be used with clock to pass the iteration to the
function instead of just repeating the same call over and over.

There must be a better, less hacky way to do this, but since clock won't
get used in real code or libraries its not high on my todo list.